st works, new driver, lots of usb stuff, more X correctness, dmenu (broken)

+++ include/X11/Xutil.h
@@ -4,7 +4,7 @@
 #include <X11/Xlib.h>
 
 typedef struct {
-  int flags;
+  long flags;
   int x, y;
   int width, height;
   int min_width, min_height;
